requirements.txt
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
pytest>=7.4.0
SQLAlchemy>=2.0.25
psycopg2-binary>=2.9.9
openai>=1.40.0
tenacity>=8.2.0
httpx>=0.27.0
google-api-python-client>=2.130.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
